requirements.txt
attrs
dj-database-url
Django
django-environ
djangorestframework
drf-spectacular
gunicorn
inflection
jsonschema
jsonschema-specifications
Pillow
psycopg
pytz
PyYAML
referencing
rpds-py
uritemplate
whitenoise
